refactor(main): extract electron-free data-paths module (unblock hooks child build)
